Move ConfigManager and ConfigSnapshot types to types.go
