+++ internal/client/config/manager.go
@@ -11,19 +11,6 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-// ConfigManager manages all configuration state in a thread-safe manner
-type ConfigManager struct {
-	mu                 sync.RWMutex
-	token              string
-	argsAttackInstance ArgsAttack
-	localConfig        ConfigLocal
-	sharedConfig       models.ConfigShared
-	useTUI             bool
-	pid                int
-	exploitName        string
-	useBanner          bool
-}
-
 var (
 	instance *ConfigManager
 	once     sync.Once
@@ -356,33 +343,4 @@ func (cm *ConfigManager) MapPortToService(port uint16) string {
 		}
 	}
 	return ""
-}
-
-// GetAllConfig returns a snapshot of all configuration data (useful for debugging)
-func (cm *ConfigManager) GetAllConfig() ConfigSnapshot {
-	cm.mu.RLock()
-	defer cm.mu.RUnlock()
-
-	return ConfigSnapshot{
-		Token:              cm.token,
-		ArgsAttackInstance: cm.argsAttackInstance,
-		LocalConfig:        cm.localConfig,
-		SharedConfig:       cm.sharedConfig,
-		UseTUI:             cm.useTUI,
-		PID:                cm.pid,
-		ExploitName:        cm.exploitName,
-		UseBanner:          cm.useBanner,
-	}
-}
-
-// ConfigSnapshot represents a point-in-time snapshot of all configuration
-type ConfigSnapshot struct {
-	Token              string
-	ArgsAttackInstance ArgsAttack
-	LocalConfig        ConfigLocal
-	SharedConfig       models.ConfigShared
-	UseTUI             bool
-	PID                int
-	ExploitName        string
-	UseBanner          bool
 }
